renamed files and added folders, created prop object for future chart components
