+++ components/Charts/CategoryChart.js
@@ -1,0 +1,3 @@
+export const CategoryChart = () => {
+  return <div>hi from cTGALKSDFJASLDK</div>;
+};
